requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
streamlit
sentencepiece
langchain_ollama